// src/models/generated/searchBlock.ts
/**
 * Ergo blockchain explorer
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 2.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */ import { Miner } from './miner';
import { ModifierId } from './modifierId';
import { Timestamp } from './timestamp';

/**
 * Block for search results
 */
export interface SearchBlock {
  id?: ModifierId;
  height: number;
  timestamp: Timestamp;
  transactionsCount: number;
  miner: Miner;
  /**
   * Miner Reward
   */
  minerReward: number;
  difficulty?: number;
  /**
   * Size of block in bytes
   */
  size: number;
}